feat(geoutils): convert singleton object to functions
