+++ src/index.ts
@@ -1,2 +1,14 @@
+export {
+  getAdministrativeRegions,
+  getAdministrativeRegionById,
+  getAdministrativeRegionByIsoCode,
+  getAdministrativeUnits,
+  getAdministrativeUnitById,
+  getMunicipalities,
+  getGeographicRegions,
+  getGeographicRegionById,
+  getPrefectures,
+  getPrefectureById,
+  findByPostalCode,
+} from "./geoUtils";
 export { getDays, getMonths, getQuarters, getEras } from "./dateUtils";
-export { GeoUtils } from "./GeoUtils";
